+++ source/js/tools/getBrowser.ts
@@ -1,6 +1,6 @@
 export const getBrowser = () => /^moz-extension:/.test(location.href) ? {
 	name: 'Firefox',
-	type:'firefox',
+	type: 'firefox',
 	pluginsUrl: 'https://addons.mozilla.org/firefox/addon/proton-vpn-firefox-extension',
 } : {
 	name: 'Chrome',